+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0]
 ### Added
 - Show error message when converting with `patternProperties`
 - Include definitions when converting to JSONoid objects
@@ -157,7 +159,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-05-31
 - Initial release
 
-[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.5...HEAD
+[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.5...v0.6.0
 [0.5.5]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.4...v0.5.5
 [0.5.4]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.2...v0.5.3